# api_docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "KrishiSetu AI Backend",
    "description": "Smart India Hackathon 2026 (SIH 26033) - Feature Phone First Farmer Marketplace",
    "version": "1.0.0"
  },
  "paths": {
...
    "/listing/create": {
      "post": {
        "tags": [
          "Marketplace Listings"
        ],
        "summary": "Create Listing",
        "operationId": "create_listing_listing_create_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ListingCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListingResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/listing/all": {
      "get": {
        "tags": [
          "Marketplace Listings"
        ],
        "summary": "Get All Listings",
        "operationId": "get_all_listings_listing_all_get",
        "parameters": [
          {
            "name": "crop",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Crop"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "state",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ListingResponse"
                  },
                  "title": "Response Get All Listings Listing All Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/listing/{id}": {
      "get": {
        "tags": [
          "Marketplace Listings"
        ],
        "summary": "Get Listing By Id",
        "operationId": "get_listing_by_id_listing__id__get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListingResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/negotiation/start": {
      "post": {
        "tags": [
          "AI Negotiation"
        ],
        "summary": "Start Negotiation",
        "description": "Starts an AI-managed negotiation round on behalf of the farmer.",
        "operationId": "start_negotiation_negotiation_start_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NegotiationStartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NegotiationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/negotiation/respond": {
      "post": {
        "tags": [
          "AI Negotiation"
        ],
        "summary": "Respond Negotiation",
        "description": "Buyer counter-offers, accepts, or rejects. AI counter-acts protecting farmer reserve price.",
        "operationId": "respond_negotiation_negotiation_respond_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NegotiationRespondRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NegotiationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/negotiation/{id}/history": {
      "get": {
        "tags": [
          "AI Negotiation"
        ],
        "summary": "Get Negotiation History",
        "operationId": "get_negotiation_history_negotiation__id__history_get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NegotiationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/mandi/crop/{crop_name}": {
      "get": {
        "tags": [
          "Mandi Intelligence"
        ],
        "summary": "Get Crop Mandi Details",
        "operationId": "get_crop_mandi_details_mandi_crop__crop_name__get",
        "parameters": [
          {
            "name": "crop_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Crop Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MandiPriceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/logistics/routes": {
      "get": {
        "tags": [
          "Pooled Logistics"
        ],
        "summary": "Get All Routes",
        "description": "Returns nearby trucks travelling along route with available capacity,\nshared cost, solo cost, and estimated savings.",
        "operationId": "get_all_routes_logistics_routes_get",
        "parameters": [
          {
            "name": "village",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "Murthal",
              "title": "Village"
            }
          },
          {
            "name": "district",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "Sonipat",
              "title": "District"
            }
          },
          {
            "name": "destination",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "Azadpur Mandi, Delhi",
              "title": "Destination"
            }
          },
          {
            "name": "quantity_kg",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "default": 200.0,
              "title": "Quantity Kg"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TruckRouteResponse"
                  },
                  "title": "Response Get All Routes Logistics Routes Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/dashboard/buyer": {
      "get": {
        "tags": [
          "Dashboards"
        ],
        "summary": "Get Buyer Dashboard",
        "operationId": "get_buyer_dashboard_dashboard_buyer_get",
        "parameters": [
          {
            "name": "buyer_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "BUY-001",
              "title": "Buyer Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ListingCreate": {
        "properties": {
          "farmer_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Farmer Id",
            "default": "FARM-001"
          },
          "farmer_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Farmer Name",
            "default": "Rameshwar Singh"
          },
          "farmer_phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Farmer Phone",
            "default": "+919812345001"
          },
          "crop_name": {
            "type": "string",
            "title": "Crop Name"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category",
            "default": "Vegetable"
          },
          "quantity_kg": {
            "type": "number",
            "title": "Quantity Kg"
          },
          "expected_price_per_kg": {
            "type": "number",
            "title": "Expected Price Per Kg"
          },
          "farmer_min_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Farmer Min Price"
          },
          "village": {
            "type": "string",
            "title": "Village"
          },
          "district": {
            "type": "string",
            "title": "District"
          },
          "state": {
            "type": "string",
            "title": "State"
          },
          "harvest_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Harvest Date",
            "default": "2026-09-06"
          },
          "quality_grade": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quality Grade",
            "default": "Grade A"
          },
          "created_via": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created Via",
            "default": "AI_HOTLINE_VOICE"
          }
        },
        "type": "object",
        "required": [
          "crop_name",
          "quantity_kg",
          "expected_price_per_kg",
          "village",
          "district",
          "state"
        ],
        "title": "ListingCreate"
      },
      "ListingResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "farmer_id": {
            "type": "string",
            "title": "Farmer Id"
          },
          "farmer_name": {
            "type": "string",
            "title": "Farmer Name"
          },
          "farmer_phone": {
            "type": "string",
            "title": "Farmer Phone"
          },
          "crop_name": {
            "type": "string",
            "title": "Crop Name"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "quantity_kg": {
            "type": "number",
            "title": "Quantity Kg"
          },
          "expected_price_per_kg": {
            "type": "number",
            "title": "Expected Price Per Kg"
          },
          "farmer_min_price": {
            "type": "number",
            "title": "Farmer Min Price"
          },
          "current_best_offer": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Best Offer"
          },
          "village": {
            "type": "string",
            "title": "Village"
          },
          "district": {
            "type": "string",
            "title": "District"
          },
          "state": {
            "type": "string",
            "title": "State"
          },
          "harvest_date": {
            "type": "string",
            "title": "Harvest Date"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "default": ""
          },
          "quality_grade": {
            "type": "string",
            "title": "Quality Grade"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "ai_mandi_benchmark": {
            "type": "number",
            "title": "Ai Mandi Benchmark"
          },
          "recommended_buyer_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recommended Buyer Id"
          },
          "created_via": {
            "type": "string",
            "title": "Created Via"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "id",
          "farmer_id",
          "farmer_name",
          "farmer_phone",
          "crop_name",
          "category",
          "quantity_kg",
          "expected_price_per_kg",
          "farmer_min_price",
          "village",
          "district",
          "state",
          "harvest_date",
          "quality_grade",
          "status",
          "ai_mandi_benchmark",
          "created_via"
        ],
        "title": "ListingResponse"
      },
      "MandiMarketItem": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "modal": {
            "type": "number",
            "title": "Modal"
          },
          "min": {
            "type": "number",
            "title": "Min"
          },
          "max": {
            "type": "number",
            "title": "Max"
          },
          "arrival_tonnes": {
            "type": "number",
            "title": "Arrival Tonnes"
          }
        },
        "type": "object",
        "required": [
          "name",
          "modal",
          "min",
          "max",
          "arrival_tonnes"
        ],
        "title": "MandiMarketItem"
      },
      "MandiPriceResponse": {
        "properties": {
          "crop": {
            "type": "string",
            "title": "Crop"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "unit": {
            "type": "string",
            "title": "Unit",
            "default": "\u20b9/kg"
          },
          "modal_price": {
            "type": "number",
            "title": "Modal Price"
          },
          "min_price": {
            "type": "number",
            "title": "Min Price"
          },
          "max_price": {
            "type": "number",
            "title": "Max Price"
          },
          "trend": {
            "type": "string",
            "title": "Trend"
          },
          "best_mandi": {
            "type": "string",
            "title": "Best Mandi"
          },
          "mandis": {
            "items": {
              "$ref": "#/components/schemas/MandiMarketItem"
            },
            "type": "array",
            "title": "Mandis",
            "default": []
          },
          "history_7d": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "History 7D",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "crop",
          "category",
          "modal_price",
          "min_price",
          "max_price",
          "trend",
          "best_mandi"
        ],
        "title": "MandiPriceResponse"
      },
      "NegotiationRespondRequest": {
        "properties": {
          "negotiation_id": {
            "type": "string",
            "title": "Negotiation Id"
          },
          "buyer_counter_offer": {
            "type": "number",
            "title": "Buyer Counter Offer"
          },
          "action": {
            "type": "string",
            "title": "Action",
            "default": "COUNTER"
          }
        },
        "type": "object",
        "required": [
          "negotiation_id",
          "buyer_counter_offer"
        ],
        "title": "NegotiationRespondRequest"
      },
      "NegotiationResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "listing_id": {
            "type": "string",
            "title": "Listing Id"
          },
          "buyer_id": {
            "type": "string",
            "title": "Buyer Id"
          },
          "buyer_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Buyer Name",
            "default": ""
          },
          "crop_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Crop Name",
            "default": ""
          },
          "quantity_kg": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quantity Kg",
            "default": 0.0
          },
          "farmer_min_price": {
            "type": "number",
            "title": "Farmer Min Price"
          },
          "current_buyer_offer": {
            "type": "number",
            "title": "Current Buyer Offer"
          },
          "current_ai_counter": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Ai Counter"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "ai_reasoning": {
            "type": "string",
            "title": "Ai Reasoning"
          },
          "offers": {
            "items": {
              "$ref": "#/components/schemas/OfferHistoryItem"
            },
            "type": "array",
            "title": "Offers",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "listing_id",
          "buyer_id",
          "farmer_min_price",
          "current_buyer_offer",
          "status",
          "ai_reasoning"
        ],
        "title": "NegotiationResponse"
      },
      "NegotiationStartRequest": {
        "properties": {
          "listing_id": {
            "type": "string",
            "title": "Listing Id"
          },
          "buyer_id": {
            "type": "string",
            "title": "Buyer Id"
          },
          "buyer_offer_per_kg": {
            "type": "number",
            "title": "Buyer Offer Per Kg"
          }
        },
        "type": "object",
        "required": [
          "listing_id",
          "buyer_id",
          "buyer_offer_per_kg"
        ],
        "title": "NegotiationStartRequest"
      },
...
      "OfferHistoryItem": {
        "properties": {
          "sender": {
            "type": "string",
            "title": "Sender"
          },
          "amount_per_kg": {
            "type": "number",
            "title": "Amount Per Kg"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "sender",
          "amount_per_kg",
          "message",
          "timestamp"
        ],
        "title": "OfferHistoryItem"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}